django-backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
flake8==7.0.0
